Move pure tf MLP -> model file
